# Add a method for marshaling directly into a user-provided buffer.